requirements.txt
# Backend Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
Werkzeug==2.3.7

# Computer Vision and AI
ultralytics==8.2.52
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
torch>=2.0.0
torchvision>=0.15.0

# File handling and utilities
python-multipart==0.0.6
python-dotenv==1.0.0
uuid==1.30

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
